Fixes to have a working environment
The base box is now build on top of Ubuntu Xenial instead of Trusty,
since newer versions of Python Tornado module will not working on older
versions of Python due lack of updated SSL/TLS version.
Provisioning of primary and secondary now allow the configuration of
keyboard mapping from the Vagrantfile.
Broken down the build-primary.sh script to three scripts, to allow
setting up a Python Virtualenv and installing all required modules over
there. This gives more freedom to update modules without risking the
Ubuntu OS and also allows to update the pip program (which had broken to
install newer modules due too old version).
The mosquitto Python module is not available in Pypi repository anymore
and the project seems to be discontinued. It was replaced with
paho-mqtt, but I'm not sure the compartibility between the modules
interface, so something can be broken while evaluating RabbitMQ MQTT
support.
Replaced Upstart scripts related to running services with the
corresponding configuration for Systemd. The old Upstart scripts were
not working on Ubuntu Xenial.
Small improvements to code readability.